backend/requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
